[Chore] #22 - Category 컴포넌트 파일 분리
